requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
colorama==0.4.6
Django==4.2.1
django-factory-boy==1.0.0
django-faker==0.2
exceptiongroup==1.1.1
factory-boy==3.2.1
fake-factory==9999.9.9
Faker==18.9.0
h11==0.14.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
outcome==1.2.0
packaging==23.1
pluggy==1.0.0
psycopg2==2.9.6
pycparser==2.21
PySocks==1.7.1
pytest==7.3.1
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python-dotenv==1.0.0
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tomli==2.0.1
trio==0.22.0
trio-websocket==0.10.2
typing-extensions==4.6.2
tzdata==2023.3
urllib3==2.0.2
wsproto==1.2.0
